headscale-config-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyyaml==6.0.2
docker==7.1.0
python-multipart==0.0.12
